refactor(router): split tag router code into router.go, cache.go, and match.go
